RAG_webscrapper/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
chromadb
bs4
python-dotenv
langchain-chroma